chore: Ajusta nome dos arquivos referente as telas de atividade
