+++ espectra-front/src/routes/Routes.jsx
@@ -1,16 +1,16 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import TelaInicial from "../pages/TelaInicial";
 import TelaCadastro from "../pages/TelaCadastro";
+import TelaEditarAtividade from "../pages/TelaEditarAtividade";
 import TelaCadastroAtividade from "../pages/TelaCadastroAtividade";
-import TelaCadastroAtividadePersonalizada from "../pages/TelaCadastroAtividadePersonalizada";
 
 function Rotas() {
   return (
     <Routes>
       <Route path="/" element={<TelaInicial />} />
       <Route path="/cadastro" element={<TelaCadastro />} />
-      <Route path="/cadastro-atividade" element={<TelaCadastroAtividade />} />
-      <Route path="/cadastro-atividade-personalizada" element={<TelaCadastroAtividadePersonalizada />} />
+      <Route path="/atividade/editar" element={<TelaEditarAtividade />} />
+      <Route path="/atividade/cadastrar" element={<TelaCadastroAtividade />} />
     </Routes>
   );
 }
